Move methods out FileUtils when only used once
