+++ core/src/main/java/org/frankframework/util/flow/FlowDiagramManager.java
@@ -23,12 +23,13 @@ import java.io.InputStream;
 import java.nio.file.Files;
 import java.util.List;
 
+import lombok.Setter;
+import lombok.extern.log4j.Log4j2;
 import org.apache.commons.lang3.StringUtils;
 import org.frankframework.configuration.Configuration;
 import org.frankframework.core.IAdapter;
 import org.frankframework.util.AppConstants;
 import org.frankframework.util.ClassUtils;
-import org.frankframework.util.FileUtils;
 import org.frankframework.util.SpringUtils;
 import org.frankframework.util.XmlUtils;
 import org.springframework.beans.BeanInstantiationException;
@@ -40,9 +41,6 @@ import org.springframework.context.ApplicationContext;
 import org.springframework.context.ApplicationContextAware;
 import org.springframework.http.MediaType;
 import org.springframework.util.MimeType;
-
-import lombok.Setter;
-import lombok.extern.log4j.Log4j2;
 
 /**
  * Utility class to generate the flow diagram for an adapter or a configuration.
@@ -79,7 +77,7 @@ public class FlowDiagramManager implements ApplicationContextAware, Initializing
 	}
 
 	/**
-	 * Optional IFlowGenerator. If non present the FlowDiagramManager should still be
+	 * Optional IFlowGenerator. If not present, the FlowDiagramManager should still be
 	 * able to generate dot files and return the `noImageAvailable` image.
 	 */
 	protected IFlowGenerator createFlowGenerator(String generatorBeanClass) {
@@ -216,7 +214,7 @@ public class FlowDiagramManager implements ApplicationContextAware, Initializing
 			throw new IllegalStateException(parent.getPath() + " does not exist and could not be created");
 		}
 
-		String name = FileUtils.encodeFileName(fileName) + "." + flowGenerator.getFileExtension();
+		String name = encodeFileName(fileName) + "." + flowGenerator.getFileExtension();
 		log.trace("retrieve flow file for name[{}] in folder[{}]", ()->fileName, parent::getPath);
 
 		return new File(parent, name);
@@ -245,6 +243,28 @@ public class FlowDiagramManager implements ApplicationContextAware, Initializing
 		}
 
 		log.debug("finished generating flow diagram for [{}] in [{}] ms", () -> name, () -> System.currentTimeMillis() - start);
+	}
+
+	static String encodeFileName(String fileName) {
+		String mark = "-_.+=";
+		StringBuilder encodedFileName = new StringBuilder();
+		int len = fileName.length();
+		for (int i = 0; i < len; i++) {
+			char c = fileName.charAt(i);
+			if ((c >= '0' && c <= '9')
+					|| (c >= 'a' && c <= 'z')
+					|| (c >= 'A' && c <= 'Z'))
+				encodedFileName.append(c);
+			else {
+				int imark = mark.indexOf(c);
+				if (imark >= 0) {
+					encodedFileName.append(c);
+				} else {
+					encodedFileName.append('_');
+				}
+			}
+		}
+		return encodedFileName.toString();
 	}
 
 	@Override
