+++ filesystem/src/test/java/org/frankframework/util/FileUtilsTest.java
@@ -5,20 +5,17 @@ import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertNull;
 import static org.junit.jupiter.api.Assertions.assertTrue;
-import static org.junit.jupiter.api.Assumptions.assumeTrue;
 
 import java.io.File;
 import java.net.URL;
 import java.nio.file.Files;
 import java.nio.file.Path;
-import java.util.Date;
 
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.io.TempDir;
 
 import org.frankframework.filesystem.FileNotFoundException;
-import org.frankframework.testutil.TestAssertions;
 
 public class FileUtilsTest {
 
@@ -67,7 +64,8 @@ public class FileUtilsTest {
 		toBeMoved.createNewFile(); //Make sure it exists
 		assertTrue(toBeMoved.exists());
 
-		String result = FileUtils.moveFile(toBeMoved, testFolderPath, true, 0);
+		File dstFile = new File(testFolderPath, toBeMoved.getName());
+		String result = FileUtils.moveFile(toBeMoved, dstFile, true, 0, 5, 500);
 		File file = new File(testFolderPath, "movingFile.txt");
 		assertTrue(file.exists());
 
@@ -79,15 +77,6 @@ public class FileUtilsTest {
 		File sourceFile = getFile("fileToAppend.txt");
 		File destFile = getFile("copyFile.txt");
     	assertTrue(FileUtils.copyFile(sourceFile, destFile, true));
-	}
-
-	@Test
-	void createTempDirectoryTest() throws Exception {
-		File f = new File(testFolderPath);
-		File file = FileUtils.createTempDirectory(f);
-		boolean b = file.exists();
-		file.delete();
-		assertTrue(b);
 	}
 
 	@Test
@@ -137,47 +126,6 @@ public class FileUtilsTest {
 		assertEquals(2, containsBothFiles);
 	}
 
-	@Test //retrieve the first file from a directory. Alphabetically it should first return 'copyFile'. Add a stability period, and check if it skips the first file
-	void testGetFirstFile() throws Exception {
-		assumeTrue(TestAssertions.isTestRunningOnWindows());
-
-		long stabilityPeriod = 5000;
-		getFile("copyFrom.txt").setLastModified(new Date().getTime()-(stabilityPeriod + 500)); //mark file as stable (add 500ms to the stability period because of HDD latency)
-		getFile("copyFile.txt").setLastModified(new Date().getTime()); //update last modified to now, so it fails the stability period
-
-		File directory = getFile(null);
-		File file = FileUtils.getFirstFile(directory);
-		assertEquals("copyFile.txt", file.getName());
-	}
-
-	@Test
-	void testAlignForValLengthLeftAlignFillchar() {
-		String s = "test";
-		String res1 = FileUtils.align(s, 10, true, 'b');
-		String res2 = FileUtils.align(s, 2, true, 'b');
-		String res3 = FileUtils.align(s, 4, false, 'b');
-		assertEquals("testbbbbbb", res1);
-		assertEquals("te", res2);
-		assertEquals("test", res3);
-	}
-
-	@Test
-	void testAlignForValLengthRightAlignFillchar() {
-		String s = "test";
-		String res1 = FileUtils.align(s, 10, false, 'c');
-		String res2 = FileUtils.align(s, 2, false, 'c');
-		String res3 = FileUtils.align(s, 4, false, 'c');
-		assertEquals("cccccctest", res1);
-		assertEquals("te", res2);
-		assertEquals("test", res3);
-	}
-
-	@Test
-	void testGetFilledArray() {
-		char[] arr = FileUtils.getFilledArray(5, 'a');
-		assertEquals("aaaaa", new String(arr));
-	}
-
 	@Test
 	void testGetFileNameExtension() {
 		String ext = FileUtils.getFileNameExtension("file.blaaaathowdiaa");
@@ -200,11 +148,6 @@ public class FileUtilsTest {
 	void testGetBaseNameWithoutExtension() {
 		String name = FileUtils.getBaseName("file-blaaaathowdiaa");
 		assertNull(name);
-	}
-
-	@Test
-	void testEncodeFileName() {
-		assertEquals("_ab__5__c.txt", FileUtils.encodeFileName(" ab&@5*(c.txt"));
 	}
 
 }
