+++ nn-specials/src/test/resources/GetFirstFile/copyFile.txt
@@ -1,0 +1,1 @@
+testFile
