+++ nn-specials/src/test/resources/GetFirstFile/copyFrom.txt
@@ -1,0 +1,1 @@
+testfile that isn't used but is alphabetically sorted after copyFile.txt
