+++ nn-specials/src/test/resources/GetFirstFile/file.txt
@@ -1,0 +1,1 @@
+will be copied here
